update make generate and fix comment issues
