+++ pkg/apis/kubeflow/v2beta1/openapi_generated.go
@@ -276,13 +276,6 @@ func schema_pkg_apis_kubeflow_v2beta1_MPIJobSpec(ref common.ReferenceCallback) c
 							Format:      "int32",
 						},
 					},
-					"launcherCreationPolicy": {
-						SchemaProps: spec.SchemaProps{
-							Description: "launcherCreationPolicy if WaitForWorkersReady, the launcher is created only after all workers are in Ready state",
-							Type:        []string{"string"},
-							Format:      "",
-						},
-					},
 					"runPolicy": {
 						SchemaProps: spec.SchemaProps{
 							Description: "RunPolicy encapsulates various runtime policies of the job.",
@@ -307,6 +300,13 @@ func schema_pkg_apis_kubeflow_v2beta1_MPIJobSpec(ref common.ReferenceCallback) c
 					"sshAuthMountPath": {
 						SchemaProps: spec.SchemaProps{
 							Description: "SSHAuthMountPath is the directory where SSH keys are mounted. Defaults to \"/root/.ssh\".",
+							Type:        []string{"string"},
+							Format:      "",
+						},
+					},
+					"launcherCreationPolicy": {
+						SchemaProps: spec.SchemaProps{
+							Description: "launcherCreationPolicy if WaitForWorkersReady, the launcher is created only after all workers are in Ready state. Defaults to AtStartup.",
 							Type:        []string{"string"},
 							Format:      "",
 						},
